Completing: ProductConsumption and filtering by consumption rate

+++ HyperStoreServiceAPP/CustomModels/ProductMetadata.cs
@@ -10,6 +10,6 @@ namespace HyperStoreServiceAPP.CustomModels
     public class ProductMetadata
     {
         public IRange<decimal?> DiscountPerRange { get; set; }
-        public IRange<decimal> QuantityRange { get; set; }
+        public IRange<float> QuantityRange { get; set; }
     }
 }
