+++ HyperStoreServiceAPP/Models/Product.cs
@@ -37,7 +37,7 @@ namespace HyperStoreService.Models
         public decimal? SGSTPer { get; set; }
         [Required]
         public decimal? Threshold { get; set; }
-        public decimal TotalQuantity { get; set; }
+        public float TotalQuantity { get; set; }
 
         public Guid? LatestSupplierId { get; set; }
         [ForeignKey("LatestSupplierId")]
